Fix the error of GitLab CI on commit ccd7cb6601e2df5084aef7e90bf4b299ea852b8c
From CockroachDB v20.2.0, a CockroachDB node started with "cockroach start"
without the "--join" flag no longer automatically initializes the cluster.
The "cockroach init" command is now mandatory. The auto-initialization
behavior had been deprecated in version 19.2.

So, added the "--join" flag to the first node, and added the "cockroach init"
command that initializes the cluster, into the process of CI.